requirements.txt
gpiozero
sparkfun-qwiic-buzzer
smbus2